Draw combo lineedit with combo frame and...
This is a big commit:

(1) Combo lineedits are drawn in drawComplexControl() instead of drawPrimitive(). The main reason was the state change animation but famous style plugins do the same thing -- although, apparently, they draw combo lineedits twice!

(2) The restrictions on combo and spin box animations are removed.

(3) The focus-in animation includes scroll areas too.

(3) A focus-out animation is added for lineedits, combos, spinboxes and scroll areas. It is independent of the focus-in animation.

(4) Also Qt-5.7 compiler warnings are quietened.